backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.4
Pillow==11.1.0
python-multipart==0.0.20
